% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{aggregate_diagonals}
\alias{aggregate_diagonals}
\title{Aggregates diagonals of based on similarity of XY distribution}
\usage{
aggregate_diagonals(
  mtx1.sparse,
  mtx2.sparse,
  agg.diags = TRUE,
  which.test = c("energy", "KS")[1],
  alpha = 0.05,
  exclude.outliers = FALSE
)
}
\arguments{
\item{mtx1.sparse}{data frame, contact map in sparse format}

\item{mtx2.sparse}{data frame, contact map in sparse format}

\item{agg.diags}{logical if false leave each diagonal in separate group (no pooling)}

\item{alpha}{numeric significance threshold for chi square test}
}
\description{
Merges 2 contact maps in sparse format and then aggregates it diagonals. The aggregation process is performed as follows:
\itemize{
  \item{}{start with k = 1 (first diagonal), group k consists of observations: val.x * val.y (from k-th diagonal)}
  \item{}{take merge candidate i.e.: k' = k + 1, , group k' consists of observations: val.x * val.y (from k'-th diagonal)}
  \item{}{perform chi-square test between group k and k', H0: observations from both groups come from the same distribution, H1: they come from different distributions}
  \item{}{if rejected set k = k' (and pool diagonals k to k'-1 into single group), otherwise set k' = k' + 1}
  \item{}{repeat until all diagonals are processed}
}
After algorithm is finished diagonals are assigned to groups (pools).
}
\examples{

}
