% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{artificial_lri}
\alias{artificial_lri}
\title{Produce artificial long range interactions}
\usage{
artificial_lri(mtx.sparse, interactions)
}
\arguments{
\item{mtx.sparse}{data frame - Hi-C contact map in sparse format with columns: i, j, val, diagonal, name}

\item{interactions}{data frame with i, j, N columns indicating where (\code{i}, \code{j}) to make artificial LRI and with how many interactions (\code{N})}
}
\value{
data frame with the same columns as \code{mtx.sparse}
}
\description{
Produces artificial long range interactions by randomly moving contacts to given interacting regions.
}
\examples{
npz <- read_npz("/home/rafal/data/hic-normalization-40kb/raw/IMR90-MboI-1_40kb-raw.npz", mtx.names = c("18"))
mtx <- npz[["18"]]
# create artificial LRI data frame:
interactions <- data.frame(i = c(4, 150, 386, 370, 463, 472), j = c(8, 180, 500, 390, 481, 485), N = c(30, 40, 100, 75, 40, 88))
# produce contact map with artificial LRI
mtx.r <- artificial_lri(mtx, interactions)
m1 <- sparse2dense(mtx, N = 1952)
m2 <- sparse2dense(mtx.r, N = 1952)
md <- m1 - m2
# see artificial LRI
plot_diff_map(md, na.color = "white", breaks = 100)

}
\seealso{
\code{\link{move_contacts}} on how single artificial LRI (\code{i},\code{j}) is created
}
