% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{balance}
\alias{balance}
\title{Inserts 0 columns and rows after last row/column to symmetrize matrix.}
\usage{
balance(mtx, N = NULL)
}
\arguments{
\item{mtx}{matrix in dense format to be symmetrized}

\item{N}{positive integer; additional argument for symmetrizing matrix to desired N x N dimension; N need not be larger than \code{ncol(mtx)} or \code{nrow(mtx)} in which case submatrix \code{mtx[1:N,1:N]} will be extraced}
}
\value{
N by N matrix which is either submatrix of \code{mtx} or \code{mtx} extended with 0's row and/or columns
}
\description{
Inserts 0 columns and rows after last row/column to symmetrize matrix.
}
\examples{
mtx1 <- matrix(1:24, ncol = 4)
mtx2 <- matrix(1:24, nrow = 4)
print(mtx1)
print(mtx2)
balance(mtx1)
balance(mtx2)
balance(mtx1, N = 8)
balance(mtx1, N = 3)

}
