% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{bootstrap_sparse}
\alias{bootstrap_sparse}
\title{Bootstraps interactions from contact map given in sparse format.}
\usage{
bootstrap_sparse(sparse.mtx, ratio = c(0.5, 0.5),
  with.replacement = FALSE)
}
\arguments{
\item{sparse.mtx}{data.frame Hi-C contact map in sparse format with mandatory columns i, j, val}

\item{ratio}{numeric vector indicating on how many atomic interaction sets should initial atomic interaction set be divided; the values inside this vector depend on which sampling is used (with or without replacement); each entry of ratio vector contains fraction of interaction to be put in corresponding atomic interactions set; ratio vector must sum to 1 and all its entries must be larger than one}

\item{with.replacement}{logical indicating whether to perform sampling with or without (default) replacement}
}
\value{
list with data frames (Hi-C maps in sparse format) containing sampled interactions (according to specified ratio vector)
}
\description{
For details of bootstrapping procedure see \code{\link{bootstrap_interactions}}.
}
\examples{
sparse.mtx <- data.frame(i = c(1,3,5,7,8,9,11), j = c(7,2,1,1,3,10,9), val = c(10,8,3,1,1,2,20))
bootstrapped <- bootstrap_sparse(sparse.mtx)
print(bootstrapped)

}
\seealso{
\code{\link{bootstrap_interactions}} for details of Hi-C interactions bootstrapping procedure
}
