% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{compartment_ranges}
\alias{compartment_ranges}
\title{Calculates ranges of each consecutive compartment along given pc vector.}
\usage{
compartment_ranges(pc)
}
\arguments{
\item{pc}{numeric, compartment vector (eigenvector)}
}
\value{
data.frame where each row corrspond to interval of consecutive same sign values of eigenvector; columns are start, end and compartment
}
\description{
Entries with the same sign (i.e. positive or negative) comprise the same compartment. Positives are assigned to A compartment and negatives to B compartment.
}
\examples{
# make artificial eigenvector
ev <- c(-0.3,-0.5,0.2,0.3,0.4,0.4,-0.5,0.2,0.1,0.3,-0.9,-0.7)
compartment_ranges(ev)

}
