% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{constructGLM}
\alias{constructGLM}
\title{Constructs GLM to model per-diagonal Hi-C contact dependancy}
\usage{
constructGLM(df, link, try.poisson = TRUE)
}
\arguments{
\item{df}{data frame with predictor, response, outlier columns}

\item{link}{character link to be used in GLM}

\item{try.poisson}{logical if true then use Poisson regression instead of NB anytime warning is generated when fitting NB model indicating MLE convergence issue}
}
\value{
object of class glm or MASS::glm.nb
}
\description{
Models Hi-C contacts using Negative Binomial (or Poisson when the data is underdispersed) regression. Given the fact that Hi-C data suffers from contact decay bias this method is intendent to model each diagonal separately.
}
\seealso{
\code{\link{glm}}, \code{\link{MASS::glm.nb}} to see how GLM are constructed
}
