% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dense2sparse}
\alias{dense2sparse}
\title{Converts matrix given in dense format to sparse format data frame.}
\usage{
dense2sparse(mtx, add.diagonal = TRUE, name = NULL)
}
\arguments{
\item{mtx}{matrix in dense format}

\item{add.diagonal}{logical, if true an additional column indicating diagonal of each cell will be appended to resulting data frame}

\item{name}{character, additional argument, if specified column with name will be appended to resulting data frame}
}
\value{
data.frame with columns \code{c("i","j","val")} and optionally \code{c("diagonal","name")} columns; every row of resulting dataframe corresponds to cell in given dense matrix with i-th row, j-th column and value val
}
\description{
This function only keeps non-zero cells. In case given dense matix is symmetric \code{dense2sparse} will return upper triangular part of the matrix (i.e. where rows <= columns)
}
\examples{
dense2sparse(matrix(1:24, ncol = 3))
dense2sparse(matrix(1:24, ncol = 3), name = "some.matrix")
dense2sparse(matrix(1:24, ncol = 3), add.diagonal = FALSE)
# symmetric matrix
mtx.sym <- matrix(1:25, ncol = 5)
mtx.sym <- mtx.sym + t(mtx.sym)
dense2sparse(matrix(mtx.sym))

}
