% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{differential_interactions.HiCcopula}
\alias{differential_interactions.HiCcopula}
\title{Finds significantly interacting rectangle-like regions.}
\usage{
\method{differential_interactions}{HiCcopula}(hic.copula,
  plot.models = TRUE, pval = 0.05, sig.thr.selection = c(1, 2, 3)[1],
  which.significance = c("qval", "pval")[1], cc.direction = c(4, 8)[1])
}
\arguments{
\item{hic.copula}{object of class HiCcopula}

\item{plot.models}{logical if true then plot bilinear model fit for every matrix in hic.copula object; it will plot models for depleted and enriched models separately; if you want to save this results to file open device before calling this function (see for instance \code{\link{pdf}}) and close device after function call (see \code{\link{dev.off}})}

\item{pval}{numeric, p-value cutoff to qualify interaction as significant}

\item{sig.thr.selection}{numeric, if 3 then only use bilinear model fit to establish p-value cutoff for significant interactions, if 2 then select significant interactions using only \code{pval} parameter, if 1 (default) use bilinear model, but if p-value threshold is larger than \code{pval}, use \code{pval} instead}

\item{which.significance}{character either "qval" or "pval" indicating, which of the 2 shold be used as a measure of interaction significance}

\item{cc.direction}{specifies criterium for two cells to be considered as neighbors during connected components search, for details see \code{directions} parameter of \code{\link{raster::clump}} function}
}
\value{
list with number of entries equal to hic.copula$names; each entry is a list with 2 elements: interacting.regions - data frame containing rows with rectangle like regions of significant interactions with coordinates n.cells (number of non zero cells inside rectangle), start.x, end.x, start.y, end.y, effect; connected.components list with cells comprising given connected component; connected components list is named list where each entry name is unique id, which can be mapped to row in interacting.regions (its row names)
}
\description{
This function works in 3 steps:
\itemize{
 \item{}{first it calculates differential map,}
 \item{}{then it takes negative log10 p-value vector of cells, sorts it, divides on negative and positive parts and lastly fits bilinear model to each of them}
 \item{}{finally it retains only those cells that are to the left (right) of intersection point of bilinear model in positive (negative) p-values vector and searches for connected components in both of them separately}
}
Fitting bilinear model is performed using \code{\link{best_fit_bilinear}} function, while for connected components search \code{\link{raster}} package is used. After detection of significantly interacting regions (connected components) one may further filter list to only retain those with number of non zero cells (n.cells column in interacting.regions data frame) larger than some threshold. There are 3 possible ways of selecting significant interactions (cells):
\itemize{
 \item{}{bilinear model is used to determine significance threshold and then this threshold is compared with \code{pval} parameter - if threshold is less significant than \code{pval} then threshold is substituted with \code{pval} - this is the default behaviour,}
 \item{}{only \code{pval} is used as a significance threshold, i.e. hard thresholding,}
 \item{}{only bilinear model is used to determine significance threshold (unrecomended, as it may yield non significant interactions).}
}
When using option 1 and 3 its recommended to plot the fit (enabled by default). An indication of properly determined significance threshold would be when red vertical line (the significance threshold) is located to the right side of grey vertical line.
}
\examples{
# first create HiCcopula object - see ?HiCcopula for examples
di <- differential_interactions(hic.copula)
di18 <- di[["18"]]
# if you want to plot results create pvalue map (see hicdiff function) and from that dense map for some chromosome
plot_diff_map(dense)
# plot regions having at least 10 significant cells in connected component
plot_regions(di18[di18$n.cells >= 10,2:6], pal.colors = c("blue","red"))

}
\seealso{
\code{\link{best_fit_bilinear}} for fitting bilinear model, \code{\link{raster::raster}} and \code{\link{raster::clump}} for connected components search
}
