% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{differential_interactions.HiCglm}
\alias{differential_interactions.HiCglm}
\title{Finds significantly interacting rectangle-like regions.}
\usage{
\method{differential_interactions}{HiCglm}(hic.glm, sig.map = NULL,
  plot.models = TRUE, pval = 0.05, sig.thr.selection = c(1, 2, 3)[1],
  which.significance = c("qvalue", "pvalue")[1], cc.direction = c(4,
  8)[1])
}
\arguments{
\item{hic.glm}{object of class HiCglm}

\item{plot.models}{logical if true then plot bilinear model fit for every matrix in hic.glm object; it will plot bilinear fit for E[Y | X] and E[X | Y] models; if you want to save this results to file open device before calling this function (see for instance \code{\link{pdf}}) and close device after function call (see \code{\link{dev.off}})}

\item{pval}{numeric, pvalue (or qvalue) cutoff to qualify interaction as significant}

\item{sig.thr.selection}{numeric, if 3 then only use bilinear model fit to establish p-value cutoff for significant interactions, if 2 then select significant interactions using only \code{pval} parameter, if 1 (default) use bilinear model, but if p-value threshold is larger than \code{pval}, use \code{pval} instead}

\item{which.significance}{character either "qvalue" or "pvalue" indicating, which of the 2 should be used as a measure of interaction significance}

\item{cc.direction}{specifies criterium for two cells to be considered as neighbors during connected components search, for details see \code{directions} parameter of \code{\link{raster::clump}} function}
}
\value{
list with number of entries equal to hic.glm$names; each entry is a list with 2 elements: interacting.regions - data frame containing rows with rectangle like regions of significant interactions with coordinates n.cells (number of non zero cells inside rectangle), start.x, end.x, start.y, end.y, effect; connected.components list with cells comprising given connected component; connected components list is named list where each entry name is unique id, which can be mapped to row in interacting.regions (its row names); effect column is indicating if interaction refers to Y enrichment (i.e. E[Y | X] model) or X enrichment (i.e. E[X | Y] model)
}
\description{
This function works in 3 steps:
\itemize{
 \item{}{first it calculates differential map,}
 \item{}{then it takes negative log significance (qvalue) vector of cells, sorts it and fits bilinear model}
 \item{}{finally it retains only those cells that are to the right of intersection point of bilinear model in significance vector (i.e. the most significant cells) and searches for connected components}
}
Fitting bilinear model is performed using \code{\link{best_fit_bilinear}} function, while for connected components search \code{\link{raster}} package is used. After detection of significantly interacting regions (connected components) one may further filter list to only retain those with number of non zero cells (n.cells column in interacting.regions data frame) larger than some threshold. There are 3 possible ways of selecting significant interactions (cells):
\itemize{
 \item{}{bilinear model is used to determine significance threshold and then this threshold is compared with \code{pval} parameter - if threshold is less significant than \code{pval} then threshold is substituted with \code{pval} - this is the default behaviour,}
 \item{}{only \code{pval} is used as a significance threshold, i.e. hard thresholding,}
 \item{}{only bilinear model is used to determine significance threshold (unrecomended, as it may yield non significant interactions).}
}
When using option 1 and 3 its recommended to plot the fit (enabled by default). An indication of properly determined significance threshold would be when red vertical line (the significance threshold) is located to the right side of grey vertical line.
}
\examples{

}
\seealso{
\code{\link{best_fit_bilinear}} for fitting bilinear model, \code{\link{raster::raster}} and \code{\link{raster::clump}} for connected components search
}
