% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{do_pca}
\alias{do_pca}
\title{PCA analysis of Hi-C contact maps.}
\usage{
do_pca(dense.mtx, ...)
}
\arguments{
\item{dense.mtx}{numeric matrix - Hi-C contact map}

\item{...}{optional arguments passed to \code{prcomp}}
}
\value{
PCA object returned by \code{prcomp} function.
}
\description{
Performs PCA on Hi-C contact map as described in Liebermann-Aiden et al 2009. More specifically it runs following routines on dense matrix:
\itemize{
 \item{}{removes unmappable regions (all zeros rows and columns)}
 \item{}{divides each diagonal of every cell by its corresponding mean of cells on diagonal}
 \item{}{converts matrix from 2. into PCC matrix}
 \item{}{performs PCA on such matrix}
 \item{}{fills in unmappable regions into PCA object vector/matrix components}
}
}
\examples{
# load Hi-C contact maps from npz file
# get sample npz file name
mtx.fname <- system.file("extdata", "MSC-HindIII-1_40kb-raw.npz", package = "CopulaHiC", mustWork = TRUE)
mtx.sparse.list <- read_npz(mtx.fname, sparse.format = TRUE)
# get matrix for selected chromosome
mtx <- mtx.sparse.list[["18"]]
# do PCA
pca <- do_pca(mtx)
print(pca)
# it is also possible to visualize results
pairs(pca)

}
\seealso{
\code{\link{prcomp}} for how is PCA performed, Lieberman-Aiden E. et al., 2009 "Comprehensive mapping of long-range interactions reveals folding principles of the human genome." for compartment detection in Hi-C contact maps.
}
