% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{dominating_signal.HiCcomparator}
\alias{dominating_signal.HiCcomparator}
\title{Calculates coverage or decay of Hi-C maps.}
\usage{
\method{dominating_signal}{HiCcomparator}(hic.comparator, which.signal = c("coverage", "decay")[1])
}
\arguments{
\item{hic.comparator}{object of type HiCcomparator}

\item{which.signal}{character either coverage or decay, which signal to calculate}
}
\value{
dataframe with following columns: (i, sum.contacts, mean.contacts, sd.contacts, name, dataset), which can be used to conveniently visualise coverages or decays (see examples)
}
\description{
Computes coverages or decays of every Hi-C maps in both data sets of given HiCcomparator object. Coverage is defined as sum of contacts on given bin. Decay is sum or mean of contacts for every diagonal.
}
\examples{
# first create HiCcomparator object - see ?HiCcomparator for examples
coverage <- dominating_signal(hic.comparator)
# visualise results
library("ggplot2")
ggplot(coverage, aes(x = i, y = sum.contacts, color = dataset)) +
geom_point(size = 0.5) +
 geom_smooth(alpha = 0.5) +
 facet_wrap(~ name, ncol = 1, scales = "free") +
 theme(legend.position = "bottom")
# get decay
decay <- dominating_signal(hic.comparator, which.signal = "decay")
ggplot(decay[decay$diagonal != 0,], aes(x = diagonal, y = mean.contacts, color = dataset)) +
geom_point(size = 0.5) +
 scale_x_log10() +
 scale_y_log10() +
 facet_wrap(~ name, ncol = 1, scales = "free") +
 theme(legend.position = "bottom")

}
\seealso{
\code{\link{HiCcomparator}} on how to construct HiCcomparator object
}
