% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{HiCcopula}
\alias{HiCcopula}
\title{An S3 object to represent Hi-C copula model.}
\usage{
HiCcopula(hic.comparator, diagonals = 0.12, include.zero.cells = FALSE,
  n.cores = 1)
}
\arguments{
\item{hic.comparator}{object of type HiCcomparator}

\item{diagonals}{fraction or numeric vector or character "all" which diagonals to use to fit models, by default fraction of chromsome length is used to indicate number of diagonals.}

\item{include.zero.cells}{logical, whether to include cells when merging maps (see \code{\link{merge.HiCcomparator}})}

\item{n.cores}{numeric number of cores to distribute model computations}
}
\value{
S3 object of class HiCcopula
}
\description{
Models diagonal-wise dependencies between Hi-C data sets with copulas. Model is constructed as follows:
\itemize{
  \item{}{merge maps1 with maps2}
  \item{}{for each diagonal in diagonals}
  \itemize{
    \item{}{take all points from this diagonal, such that they are non zero in map1 (X) and non zero in map2 (Y)}
    \item{}{model X with gamma distribution}
    \item{}{model Y with gamma distribution}
    \item{}{transform X and Y to U and V ~ Uniform(0,1) - see \code{\link{VineCopula::pobs}}}
    \item{}{model F(U,V) with copula (see \code{\link{VineCopula::BiCopSelect}})}
  }
}
Before fitting the model it's recommended to first inspect correlations between analyzed Hi-C maps before fixing this variable. As the ratio of noise / signal in Hi-C data increases rapidly with decay it's unadvised to use all diagonals for modelling. The number of diagonals to be used will depend on chromosome length, resolution and data quality. As a rule of thumb the number of diagonals should not exceed 0.2 times length of chromosome.
}
\examples{
# first create HiCcomparator object - see ?HiCcomparator for examples
# construct model
hic.copula <- HiCcopula(hic.comparator)
# load below packages for visualisation
library("fitdistrplus")
library("VineCopula")
# illustrate gamma fit of X and Y for chromosome 18, diagonal 5
plot(copula$model[["18"]][["5"]]$marginal.x)
plot(copula$model[["18"]][["5"]]$marginal.y)
# illustrate copula fit of U and V for chromosome 18, diagonal 5
plot(copula$model[["18"]][["5"]]$bf.copula)

}
\seealso{
\code{\link{HiCcomparator}} on how to construct HiCcomparator object, \code{\link{fitdistrplus::fitdist}} on distribution fitting, \code{\link{VineCopula::pobs}} on pseudo observations generation and \code{\link{VineCopula::BiCopSelect}} on finding optimal fit bivariate copula
}
