% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{hicdiff.HiCcopula}
\alias{hicdiff.HiCcopula}
\title{Computes differential (p-value) map.}
\usage{
\method{hicdiff}{HiCcopula}(hic.copula, marginal.distr = c("fit",
  "obs")[1])
}
\arguments{
\item{hic.copula}{object of class HiCcopula}

\item{marginal.distr}{character wither fit or obs; if fit then fitted gamma distribution for X and Y is used to convert them to U and V respectively}
}
\value{
list of data frames corresponding to Hi-C contact maps; each data frame contain columns: i, j, name, val.x, val.y, u, v, effect, p.value, p.value.corrected
}
\description{
Given HiCcopula object calculates depletion/enrichment p-values of cell along diagonals w.r.t. background copula models (separate for every diagonal, see \code{\link{HiCcopula} for details}).
}
\examples{
# first create HiCcopula object - see ?HiCcopula for examples
# calculate p-values and select chromosome of interest (18 in this example)
md <- hicdiff(hic.copula)[["18"]]
# convert corrected p-values to -log10(pvals)
md$neg.log.cor.pvals <- md$p.value.corrected
# make neg.log.cor.pvals negative for depleted cells
md[md$effect == "depletion", "neg.log.cor.pvals"] <- -md[md$effect == "depletion", "neg.log.cor.pvals"]
# convert sparse matrix to dense matrix
dense <- sparse2dense(md[c("i","j","neg.log.cor.pvals")], N = hic.copula$maps.dims[["18"]][1,1])
# plot results
plot_diff_map(dense)

}
\seealso{
\code{\link{HiCcopula}} on how to construct HiCcopula, \code{\link{maps_difference_diagonal}} and \code{\link{copula_pvals}} on how p-values are calculated
}
