% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{hicdiff2mtx}
\alias{hicdiff2mtx}
\title{Converts significance maps to dense matrices.}
\usage{
hicdiff2mtx(hicdiff.list, maps.dims, val.column = c("p.value.corrected",
  "p.value")[1], neg.log = TRUE, mark.depleted = TRUE)
}
\arguments{
\item{hicdiff.list}{list of data frames, output from \code{\link{hicdiff}} function}

\item{maps.dims}{list of data frames, usually an attribute of HiCcopula object that was used to produce hicdiff.list}

\item{val.column}{character indicating which column of sparse significance map to use as cell value for dense matrix (one of p.value or p.value.corrected)}

\item{neg.log}{logical wheter to apply -log10(.) to val.column}

\item{mark.depleted}{logical whether to mark depleted cells as negative}
}
\value{
list with dense matrices
}
\description{
Given list of significance maps in sparse format produced by \code{\link{hicdiff}} function converts them to dense matrix format.
}
\examples{
# first create HiCcopula object - see ?HiCcopula for examples
# then produce hicdiff.list
hicdiff.list <- hicdiff(hic.copula)
dense.hicdiff.list <- hicdiff2mtx(hicdiff.list)
# elements of dense.list can be visualized
plot_diff_map(dense.hicdiff.list[["18"]], breaks = 100)

}
\seealso{
\code{\link{hicdiff}} for generation of hicdiff.list
}
