% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{HiCglm}
\alias{HiCglm}
\title{An S3 object to represent differential GLM Hi-C model.}
\usage{
HiCglm(hic.comparator, diagonals = 0.1, remove.outliers = TRUE,
  outlier.weight = 0, link = c("log", "sqrt")[1])
}
\arguments{
\item{hic.comparator}{object of type HiCcomparator}

\item{diagonals}{fraction or numeric vector or character "all" which diagonals to use to fit models, by default fraction of chromsome length is used to indicate number of diagonals.}

\item{remove.outliers}{logical if true try to remove outliers before fitting proper model (NB regression) using robust regression (IRLS) with bisquare weight function}

\item{outlier.weight}{numeric weight threshold to remove outliers}

\item{link}{character either log or sqrt - link function for NB regression}
}
\value{
S3 object of class HiCglm
}
\description{
Models diagonal-wise dependencies between Hi-C data sets with GLM. Model is constructed as follows:
\itemize{
  \item{}{merge maps1 with maps2}
  \item{}{for each diagonal in diagonals}
  \itemize{
    \item{}{take all points from this diagonal, such that they are non zero in map1 (X) and non zero in map2 (Y)}
    \item{}{remove outliers using robust regression (recommended)}
    \item{}{apply transformation to predictor variable (X or Y): log or sqrt depending on link function}
    \item{}{model response (Y = f(X) or X = f(Y)) using Negative Binomial distribution with appropriate link function}
  }
}
Before fitting the model it's recommended to first inspect correlations between analyzed Hi-C maps before fixing this variable. As the ratio of noise / signal in Hi-C data increases rapidly with decay it's unadvised to use all diagonals for modelling. The number of diagonals to be used will depend on chromosome length, resolution and data quality. As a rule of thumb the number of diagonals should not exceed 0.1 times length of chromosome.
}
\examples{

}
\seealso{
\code{\link{HiCcomparator}} on how to construct HiCcomparator object and \code{\link{robustreg::robustRegBS}} on how robust regression is performed and outliers selection process
}
