% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{image_plot_na}
\alias{image_plot_na}
\title{Wrapper for fields::image.plot function.}
\usage{
image_plot_na(z, breaks, col, na.color = "black",
  neg.inf.color = "gold", pos.inf.color = "darkgreen",
  colorbar = TRUE, ...)
}
\arguments{
\item{z}{numeric matrix to be plotted; may contain -Inf, Inf and NA values}

\item{breaks}{numeric vector of breaks for colorscale}

\item{col}{character vector of hex color strings; usually generated from some color pallette; it's length must be equal to \code{length(breaks) - 1}}

\item{na.color}{color for NA values}

\item{neg.inf.color}{color for -Inf values}

\item{pos.inf.color}{color for Inf values}

\item{...}{additional arguments passed to \code{\link{fields::image.plot}}}
}
\description{
Allows to mark -Inf, Inf and NA values in heatmaps with different colors. It will automatically detect such values and assign them color.
}
\examples{
# matrix of data
mtx <- toeplitz(c(5:1))
# make lower triangle part negative
mtx[lower.tri(mtx)] <- -mtx[lower.tri(mtx)]
# make some cells -Inf
mtx[matrix(c(2,1,2,2,3,2), ncol = 2, byrow = TRUE)] <- -Inf
# make some cells Inf
mtx[matrix(c(3,5,4,5), ncol = 2, byrow = TRUE)] <- Inf
# make some cells NA
mtx[matrix(c(1,3,2,3,5,3), ncol = 2, byrow = TRUE)] <- NA
print(mtx)
# prepare breaks --> symmetric, from -5 to 5, spaced by 2, with 0 in the middle
# one can also introduce here log, sqrt or other scales by applying proper transformations
breaks <- sort(c(0,seq(-5,5,2)))
print(breaks)
# prepare symmetric color pallette indicating negtive values with blue, middle with white and positive with red
colors.pal = c("blue","white","red")
pal = colorRampPalette(colors.pal)
colors <- pal(length(breaks) - 1L)
# finally plot matrix
image_plot_na(mtx, breaks, colors)
}
\seealso{
\code{\link{fields::image.plot}} for function which finally handles heatmap plotting
}
