% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{interactions2zscores}
\alias{interactions2zscores}
\title{Convert interactions into zscores}
\usage{
interactions2zscores(mtx.sparse, substitute.val = TRUE)
}
\arguments{
\item{mtx.sparse}{data frame representing contact map}

\item{substitute.val}{logical indicating whether zscores should be substituted into val}
}
\value{
data frame with interaction zscores calculated over every diagonal separatelly
}
\description{
Given single (i.e. one chromosome) contact map in sparse format convert its interactions into zscores diagonal-wise.
}
\examples{

}
