% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{intersect_tads}
\alias{intersect_tads}
\title{Finds intersection between 2 sets of TADs.}
\usage{
intersect_tads(tads1, tads2)
}
\arguments{
\item{tads1}{data frame of TAD set 1 with columns: start, end}

\item{tads2}{data frame of TAD set 2 with columns: start, end}
}
\value{
data frame with start, end columns containing TAD intersection intervals
}
\description{
Intersection is defined for the same chromosomes, for example for 2 TAD sets from chromosome 18 if there is TAD in set1 with coordinates 100, 120 (start, end respectively) and in set2 there is a TAD with coordinates 110, 140 then their intersection is interval with coordinates: 110, 120.
}
\examples{
# simple artificial data set of TADs
tads1 <- data.frame(start = c(1,10,30), end = c(5,15,35))
tads2 <- data.frame(start = c(3,14,28), end = c(12,18,40))
tads.intersection <- intersect_tads(tads1, tads2)
print(tads.intersection)

}
\seealso{
\code{\link{intervals::Intervals}}, \code{\link{intervals::interval_intersection}} for functions used to find intersection between 2 sets of intervals
}
