% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{local.max}
\alias{local.max}
\title{Finds local maximas indices.}
\usage{
local.max(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
numeric vector with indices of local maxima elements of vector v (i.e. where elements of second order difference equals -2)
}
\description{
Seeks for local maximas in vector v using simple second order difference.
}
\examples{
v <- c(1,2,3,4,5,6,5,3,2,-1,-4,-3,-1,2,4,10,12,11,5,3)
idx <- local.max(v)
print(idx)
print(v[idx])
}
