% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{maps_difference_diagonal}
\alias{maps_difference_diagonal}
\title{Computes p-values given copula model.}
\usage{
maps_difference_diagonal(uv, copula.model, mht.correction = TRUE)
}
\arguments{
\item{uv}{matrix of dimension n x 2 with U,V r.v. ~ Uniform(0,1), see \code{\link{VineCopula::pobs}} for generation of U,V from X,Y}

\item{copula.model}{object of class \code{\link{VineCopula::BiCop}}}

\item{mht.correction}{logical whether to apply Benjamini Hochberg correction for multiple hypothesis testing since we are testing a number of hypothesis (diagonal cells) against null model}
}
\value{
data frame with columns: u, v, effect, p.value, p.value.corrected
}
\description{
Given copula model for diagonal of Hi-C contact map it calculates deviation from this model (p-values) in both directions (i.e. depletion and enrichment). It divides U,V matrix on 3 groups: U > V (enrichment), U == V (no effect), U < V (depletion) and calls \code{\link{copula_pvals}} for each group. Afterwards it merge results.
}
\examples{
# see copula_pvals function
}
\seealso{
\code{\link{copula_pvals}} for p-value calculation
}
