% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{mean_expected}
\alias{mean_expected}
\title{Constructs Toeplitz matrix with means on diagonals (calculated as arithmetic mean of diagonal).}
\usage{
mean_expected(dense.mtx)
}
\arguments{
\item{dense.mtx}{matrix in dense format}
}
\value{
numerical dense matrix where all entries on i-th diagonal contain mean of i-th diagonal
}
\description{
Constructs Toeplitz matrix with means on diagonals (calculated as arithmetic mean of diagonal).
}
\examples{
mtx1 <- toeplitz(c(1,2,3,4))
mean_expected(mtx1)
mtx2 <- matrix(1:16, ncol = 4)
mean_expected(mtx2)

}
\seealso{
\code{\link{toeplitz}} for more information about toeplitz matrices
}
