% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{merge.HiCcomparator}
\alias{merge.HiCcomparator}
\title{Finds regions, which interacts in both experiments (maps).}
\usage{
\method{merge}{HiCcomparator}(x, include.zero.cells = FALSE)
}
\arguments{
\item{x}{HiCcomparator object}

\item{include.zero.cells}{logical, whether to include cells, which have non zero number of contacts in one map, but not the other, not recommended}
}
\value{
list of data frames with merged contact maps data in sparse format
}
\description{
Merges Hi-C contact maps data frames 1 and 2 of HiCcomparator object by i, j, diagonal, name columns.
}
\examples{
# first create HiCcomparator object - see ?HiCcomparator for examples
merged <- merge(hic.comparator)

}
\seealso{
\code{\link{HiCcomparator}} on how to construct HiCcomparator object
}
