% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{MSC-HindIII-1_40kb-raw_maps}
\alias{MSC-HindIII-1_40kb-raw_maps}
\title{Npz file with sample Hi-C contact maps dataset.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52457}
}
\description{
Npz file containing Hi-C contact maps of human MSC in 40kb resolution. The data comes from Dixon et al., 2015 study and was processed by Imakaev et al., 2012 pipeline without iterative correction step (so it is raw data). It contains contact maps for chromosomes 17, 18, 19.
}
\examples{
# get file name
mtx.fname <- system.file("extdata", "MSC-HindIII-1_40kb-raw.npz", package = "CopulaHiC", mustWork = TRUE)
}
\seealso{
\url{https://docs.scipy.org/doc/numpy-1.13.0/reference/generated/numpy.savez_compressed.html} for npz format description, Dixon et al., 2015 "Chromatin architecture reorganization during stem cell differentiation" for study where this dataset comes from, Imakaev et al., 2012 "Iterative correction of Hi-C data reveals hallmarks of chromosome organization." for Iterative Correction of Hi-C contact maps and \url{https://mirnylab.bitbucket.io/hiclib/index.html} for its python implementation
}
