% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{neighborhood}
\alias{neighborhood}
\title{Extracts neighborhood around i,j cells of given matrix.}
\usage{
neighborhood(
  mtx.dense,
  k = 1,
  max.d = 0.15,
  m = 1,
  which.ngb = c("cross", "donut")[1],
  include.ij = FALSE
)
}
\arguments{
\item{mtx.dense}{numeric symmetric matrix}

\item{k}{numeric radius size}

\item{max.d}{numeric fraction of domains to be taken, i.e.: if \code{mtx.dense} is of size \code{n} (i.e. it have \code{n} diagonals) then \code{floor(max.d * n)} diagonals, starting at main diagonal will be taken and remaining will be discarded.}

\item{m}{numeric how many diagonals (starting at main diagonal) to discard, by definition of cross this must be fixed to m > k to have effect as m will be finally be selected as: \code{max(k,m)}}

\item{which.ngb}{character indicating the shape of neighborhood to extract, either "cross" or "donut"}

\item{include.ij}{logical whether to include the value of cell for which neighborhood is extracted}
}
\description{
Given matrix \code{mtx.dense} and positive integer \code{k} it extracts cross pattern cells:
\itemize{
 \item{}{i - k, j}
 \item{}{i, j - k}
 \item{}{i - (k-1), j}
 \item{}{i, j - (k-1)}
 \item{}{.}
 \item{}{.}
 \item{}{.}
 \item{}{i - 1, j}
 \item{}{i, j - 1}
 \item{}{i + 1, j}
 \item{}{i, j + 1}
 \item{}{.}
 \item{}{.}
 \item{}{.}
 \item{}{i + (k - 1), j}
 \item{}{i, j + (k - 1)}
 \item{}{i + k, j}
 \item{}{i, j + k}
 \item{}{i, j}
}
or donut shape. The above cells are columns in resulting data frame, so it has in total 2 + 4 * k + 1 columns: first two are cell coordinates in initial matrix, i.e.: i and j, remaining columns are the above values (in the order specified above). Selection of cross size \code{k} enforces that cells belonging to rows, columns or diagonals (starting at main diagonal) with indices 1 to \code{k} or \code{n-k+1} to \code{n} can't be assigned any value.
}
\examples{

}
