% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pc2mtx}
\alias{pc2mtx}
\title{Matches compartment with entries of contact map given in sparse format.}
\usage{
pc2mtx(pc, sparse.mtx)
}
\arguments{
\item{pc}{numeric, compartment vector or principal component (eigenvector)}

\item{sparse.mtx}{data.frame, matrix in sparse format containg manadatory fields i and j}
}
\value{
data.frame - contact map in sparse format with additional columns for every row (dense matrix cell): compartment.i, compartment.j, compartment
}
\description{
Compartment vector indices correspond to bins along chromosome.
}
\examples{
# make artificial eigenvector --> its length is 12 so it represents artificial chromosome with 12 bins
pc <- c(-0.3,-0.5,0.2,0.3,0.4,0.4,-0.5,0.2,0.1,0.3,-0.9,-0.7)
# make artificial sparse contact map
sparse.mtx <- data.frame(i = c(1,3,5,7,8,9,11), j = c(7,2,1,1,3,10,9), val = c(10,8,3,1,1,2,20))
print(sparse.mtx)
pc2mtx(pc, sparse.mtx)

}
