% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_copula_density}
\alias{plot_copula_density}
\title{Plots bivariate 2D density plot (heatmap).}
\usage{
plot_copula_density(data, palette = "RdBu", direction = -1)
}
\arguments{
\item{data}{data frame with columns U,V ~ Uniform(0,1)}

\item{palette}{character, Rcolor palette, by default RdBu}

\item{direction}{numeric 1 or -1 (revert color palette)}
}
\value{
ggplot object
}
\description{
Plots bivariate 2D density plot (heatmap).
}
\examples{
# see ?CopulaHiC::copula_pvals

}
\seealso{
\code{\link{CopulaHiC::copula_pvals}} for nice example of this function usage
}
