% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_pc_vector}
\alias{plot_pc_vector}
\title{Plots A/B compartment vector.}
\usage{
plot_pc_vector(pc, colors = c("blue", "red"), ...)
}
\arguments{
\item{pc}{numeric vector}
}
\description{
Plots A/B compartment vector.
}
\examples{
# for real data example check ?HiCcomparator
# below is with simulated data
v <- sin(seq(1, 10, length.out = 100))
plot_pc_vector(v)

}
\seealso{
\code{\link{do_pca}} on A/B compartments and how to determine them, \code{\link{HiCcomparator}} object on real data example of compartments and plotting them (examples section)
}
