% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_with_inset}
\alias{plot_with_inset}
\title{Plots contact map or diff map with inset.}
\usage{
plot_with_inset(
  args.map,
  xlim,
  ylim,
  which.map = c("contact.map", "diff.map")[1],
  mar = c(1, 2, 1, 0.1),
  args.regions = NULL
)
}
\arguments{
\item{args.map}{named list of arguments for map plotting function for type of args see \code{\link{plot_contact_map}} and \code{\link{plot_diff_map}}}

\item{xlim}{numeric 2-element vector o x limits}

\item{ylim}{numeric 2-element vector o y limits}

\item{which.map}{character string indicating if contact map or diff map should be plotted}

\item{mar}{numeric vector of length 4 specifying margins}

\item{args.regions}{named list of arguments passed to \code{\link{plot_regions}} function, if NULL then don not plot regions}
}
\description{
Additionally it can plot regions on both orignal image and inset.
}
\examples{
# get Hi-C map file
mtx.fname <- system.file("extdata", "MSC-HindIII-1_40kb-raw.npz", package = "DIADEM", mustWork = TRUE)
# read it and take chromosome 18
m.sparse18 <- read_npz(mtx.fname, mtx.names = c("18"))[["18"]]
dense <- sparse2dense(m.sparse18[c("i","j","val")], N = 1952)
plot_with_inset(list(dense), c(500,800), c(500,800))
# get TADs
tads <- map2tads(dense)
# plot with TADs
plot_with_inset(list(dense), c(500,800), c(500,800), args.regions = list(tads))

}
\seealso{
\code{\link{plot_contact_map}}, \code{\link{plot_diff_map}} for plotting contact maps and difference maps and \code{\link{plot_regions}} for plotting regions and its arguments
}
