% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotHiCglm}
\alias{plotHiCglm}
\title{Creates expected vs observed plot for given Hi-C GLM}
\usage{
plotHiCglm(hic.glm, diagonals = "all", N = 30, ncol = NULL)
}
\description{
Creates list of lists wtih plots. Each list contains a 2 element list for given model (usually chromosome), where 2 entries are: model Y | X and X | Y
}
