% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{proportions}
\alias{proportions}
\title{Calculate proportion of interactions}
\usage{
proportions(mtx.sparse)
}
\arguments{
\item{mtx.sparse}{data frame representing contact map}
}
\value{
data frame with interactions proportion calculated over every diagonal separatelly
}
\description{
Appends column with proportion of interactions per diagonal.
}
\examples{

}
