% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_dense}
\alias{read_dense}
\title{Reads given npz file and returns list with dense matrices.}
\usage{
read_dense(path, mtx.names = "all")
}
\arguments{
\item{path}{character, string specifying path to npz file}

\item{mtx.names}{character vector specyfing subset of matrices names to read from npz dict like file; by default all matrices are loaded}
}
\description{
This function requires python and numpy.
}
\examples{
# get sample npz file name
mtx.fname <- system.file("extdata", "MSC-HindIII-1_40kb-raw.npz", package = "CopulaHiC", mustWork = TRUE)
mtx.dense.list <- read_dense(mtx.fname) # reads all chromosomes
# limiting number of chromosomes
mtx.dense.sublist <- read_dense(mtx.fname, mtx.names = c("18","19")) # only read chromosome 18 and 19

}
\seealso{
\url{https://www.python.org/} for python, \url{https://www.numpy.org/} for numpy
}
