% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_npz}
\alias{read_npz}
\title{Reads npz file with matrices and converts them sparse matrices.}
\usage{
read_npz(path, mtx.names = "all", sparse.format = TRUE)
}
\arguments{
\item{path}{character, string specifying path to npz file}

\item{mtx.names}{character vector specyfing subset of matrices names to read from npz dict like file; by default all matrices are loaded}

\item{sparse.format}{logical if FALSE then this function is equivalent to \code{\link{read_dense}} function}
}
\value{
list with matrices in sparse format
}
\description{
Reads npz file with matrices and converts them sparse matrices.
}
\examples{
# get sample npz file name
mtx.fname <- system.file("extdata", "MSC-HindIII-1_40kb-raw.npz", package = "DIADEM", mustWork = TRUE)
mtx.sparse.list <- read_npz(mtx.fname) # reads all chromosomes
mtx.sparse.sublist <- read_npz(mtx.fname, mtx.names = c("18","19")) # only read chromosome 18 and 19

}
\seealso{
\code{\link{dense2sparse}} for conversion of dense matrix to sparse format, \code{\link{read_dense}} on reading npz files
}
