% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_size}
\alias{read_size}
\title{Reads dimension of every matrix in npz file.}
\usage{
read_size(path, mtx.names = "all")
}
\arguments{
\item{path}{character, string specifying path to npz file}

\item{mtx.names}{character vector specyfing subset of matrices names to read from npz dict like file; by default all matrices are loaded}
}
\value{
matrix where row names are names of matrices from npz file and columns are rows and cols; each cell of the matrix contains number of rows and columns respectively that contact map with given name consits of
}
\description{
Reads dimension of every matrix in npz file.
}
\examples{
# get sample npz file name
mtx.fname <- system.file("extdata", "MSC-HindIII-1_40kb-raw.npz", package = "CopulaHiC", mustWork = TRUE)
sizes <- read_size(mtx.fname)
print(sizes)

}
