% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{remove_unmappable}
\alias{remove_unmappable}
\title{Removes unmappable regions (all zeros columns and rows) from dense matrix.}
\usage{
remove_unmappable(dense.mtx)
}
\arguments{
\item{dense.mtx}{numeric matrix (contact map) in dense format}
}
\value{
list containing 3 elements: indices of removed rows, indices of removed columns and matrix without unmappable regions
}
\description{
Removes unmappable regions (all zeros columns and rows) from dense matrix.
}
\examples{
# construct matrix
mtx <- matrix(1:32, ncol = 4)
mtx[c(5,7,8),] <- 0
mtx[,c(2,4)] <- 0
l <- remove_unmappable(mtx)
print(l[["indices.rows"]])
print(l[["indices.cols"]])
print(l[["matrix"]])

}
