% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{restore_unmappable_vec}
\alias{restore_unmappable_vec}
\title{Restores deleted regions (cells) in vector, for example unmappable regions in pc vector.}
\usage{
restore_unmappable_vec(vec, idx, empty.elem = NA)
}
\arguments{
\item{vec}{numeric vector}

\item{idx}{indices of elements in initial vector (before cells removal) - the one which is to be restored}

\item{empty.elem}{numeric or NA how to fill missing (restored) cells}
}
\description{
Restores deleted regions (cells) in vector, for example unmappable regions in pc vector.
}
\examples{
# create vector with zeros
v <- c(1,2,3,0,0,0,2,2,0,9,8,0)
# get indices of 0 elements
idx <- which(v == 0)
v.without <- v[-idx]
restore_unmappable_vec(v.without, idx)
restore_unmappable_vec(v.without, idx, empty.elem = 0)

}
