% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{right.min}
\alias{right.min}
\title{Find first local minimum.}
\usage{
right.min(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
numeric value of first local minimum in \code{v} starting from the begining of v and going right; NA if v is nondecreasing starting at first element of v and going right (i.e. v is either constant or there is maximum first)
}
\description{
Seeks first local minimum starting from the left hand side of given vector and moving towards right hand side.
}
\examples{
# maximum first (in 7, index 3)
v1 <- c(5,6,7,6,4,3,2,1)
# no maximum or minimum
v2 <- c(2,2,2,2,2,2)
# minimum at 2 (index 3) and no maximum starting at 6 (index 7)
v3 <- c(4,3,2,3,4,5,6)
print(left.max(v1))
print(left.max(v2))
print(left.max(v3))
}
\seealso{
\code{\link[DIADEM]{local.min}}, \code{\link[DIADEM]{local.max}} for finding local minma and maxima in vector v
}
