% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sample_alri}
\alias{sample_alri}
\title{Randomly selects cells for artificial long range interactions.}
\usage{
sample_alri(
  mtx,
  N = 100,
  max.diag = NA,
  radius = 1,
  alri.min.dist = 2 * radius + 1
)
}
\arguments{
\item{mtx}{numeric Hi-C contact map in dense format}

\item{N}{numeric number of cells to sample}

\item{max.diag}{numeric specifies diagonal range to sample interactions from}

\item{radius}{numeric number of cells to include in artificial interaction around (left, right, bottom, top or corners) with sampled interaction}

\item{alri.min.dist}{numeric minimum distance (in number of cells left, right, bottom, top or diagonally) between any pair of sampled interactions (between their centers)}
}
\value{
matrix with id, i, j columns indicating artificial long range interactions
}
\description{
Toghether with \code{\link{DIADEM::simulate_map.HiCglm}} function this can be used to create Hi-C contact maps with artificial long range differential interactions (see examples).
}
\details{
Interaction is a square centered on i,j with height i - radius, ..., i, ..., i + radius and width j - radius, ..., j, ..., j + radius

The minmum distance between any pair of interactions \code{alri.min.dist} is measured as Euclidean distance between centers of 2 intractions (like i1, j1 and i2, j2)
}
