% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_hic_data}
\alias{sample_hic_data}
\title{Sample data containing Hi-C contact maps as well as HiCcomparator and HiCglm objects.}
\format{An object of class \code{list} of length 5.}
\usage{
data(sample_hic_data)
}
\description{
The list contains 5 elements:
\describe{
 \item{"IMR90-MboI-1"}{List of raw Hi-C contact maps in sparse format (i.e.: data frame with columns: i,j,val,diagonal,name) from GSE63525 study. The list only contains contact map of chromosome 21, first replicate.}
 \item{"MSC-HindIII-1"}{List of raw Hi-C contact maps in sparse format (i.e.: data frame with columns: i,j,val,diagonal,name) from GSE52457 study. The list only contains contact map of chromosome 21, first replicate.}
 \item{"HiCcomparator"}{HiCcomparator object constructed from the above 2 datasets with parameters as indicated in vignette.}
 \item{"HiCglm"}{HiCglm object constructed from the above 2 datasets with parameters as indicated in vignette.}
 \item{"chromosome.sizes"}{List with names matching names (chromosomes) from first two datasets and values equal to their sizes (number of bins in contact map).}
}
}
\keyword{datasets}
