% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_hic_maps}
\alias{sample_hic_maps}
\title{Sample Hi-C contact maps dataset.}
\format{list of length 4 with Hi-C contact maps of 2 cell lines, each in 2 replicates. Each entry of list is a 2 element list with data frames. Every data frame is Hi-C contact map of single chromosome in sparse format:
\describe{
  \item{i}{row (y) coordinate}
  \item{j}{column (x) coordinate}
  \item{val}{number of contact between i-th and j-th region}
  \item{diagonal}{distance between i and j, i.e. abs(i-j) and diagonal of Hi-C contact map which this cell belongs}
  \item{name}{Hi-C contact map name, chromosome in this case}
}}
\usage{
data(sample_hic_maps)
}
\description{
The list contains 4 entries - Hi-C datasets: IMR90-MboI-1_40kb-raw, IMR90-MboI-2_40kb-raw, MSC-HindIII-1_40kb-raw, MSC-HindIII-2_40kb-raw in 40kb resolution. This data was taken from 2 studies: Rao et al. 2014 "A 3D map of the human genome at kilobase resolution reveals principles of chromatin looping." (IMR90 datasetss) and Dixon 2015 et al. "Chromatin architecture reorganization during stem cell differentiation" (MSC datasets). Data was processed using ICE pipeline \url{https://mirnylab.bitbucket.io/hiclib/index.html} without iterative correction. Each entry in list is a list with Hi-C maps in sparse format (data frames). The data is truncated - it only contains chromosomes 18 and 19.
}
\keyword{datasets}
