% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tads}
\alias{sample_tads}
\title{Sample TAD dataset.}
\format{list of length 4 with TADs of 2 cell lines, each in 2 replicates. Each entry in list is a data frame with TADs of chromsome 18 and 19:
\describe{
  \item{start}{start bin of a TAD (first TAD starts at 0)}
  \item{end}{end bin of a TAD (for consecutive TADs it is equal to next TAD start bin)}
  \item{name}{name of Hi-C contact map on which this TADs were determined - chromosome in this case}
}}
\usage{
data(sample_tads)
}
\description{
The list contains a set of TADs determined on \code{\link{sampleHiCmaps}} dataset using Insulation Score with window size 1Mbp and delta window size 200Kbp
}
\keyword{datasets}
