% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{significance}
\alias{significance}
\title{Computes significance of data given the model}
\usage{
significance(dataset, model)
}
\arguments{
\item{dataset}{data frame with predictor, response, outlier columns}

\item{model}{numeric, 3 element vector: dispersion, intercept, beta}
}
\value{
numeric vector with p-values (upper tail)
}
\description{
Uses either Negative Binomial or Poisson to calculate pvalue of Y | X or X | Y.
}
\seealso{
\code{\link{pnbinom}}, \code{\link{ppois}} to see how to calculate significance
}
