% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{simulate_contacts}
\alias{simulate_contacts}
\title{Simulate corresponding interactions based on given interactions vector and model.}
\usage{
simulate_contacts(interaction.vector, model, N = 1)
}
\arguments{
\item{interaction.vector}{integer vector with interactions}

\item{model}{numeric, 3 element vector: dispersion, intercept, beta}

\item{N}{numeric number of interactions to sample for every interaction in \code{interaction.vector}}
}
\value{
data frame with columns: predictor (interaction.vector) and response (simulated corresponding interactions)
}
\description{
Simulate corresponding interactions based on given interactions vector and model.
}
\examples{

}
