% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{simulate_map.HiCglm}
\alias{simulate_map.HiCglm}
\title{Simulates maps based on background model.}
\usage{
\method{simulate_map}{HiCglm}(hic.glm, dpools = "all", remove.na.contacts = TRUE)
}
\arguments{
\item{hic.glm}{HiCglm object}

\item{dpools}{character or vector, if character "all" then use all pooled diagonals in background model, otherwise use specified subset}

\item{remove.na.contacts}{logical if true remove NAs}
}
\value{
list of list of data frames with expected interactions, first list enumerates different maps (usually chromosomes), second list contains 2 elements refering to Y (simulate Y condition from X) and X (simulate X condition from Y) contact maps in sparse format.
}
\description{
This function can be used to prepare artificial replicate maps (with no differential interactions) from any 2 Hi-C contact maps. To do so one must first prepare HiCglm object from any pair of contact maps (can be technical or biological replicates) and then feed it into this function. Resulting maps can be used to benchmark DI detection methods (for example by adding some artificial DI and then testing it). This approach has an advantage of producing artificial maps resembling real Hi-C data, because it keeps decay, compartment and TADs unaffected (at least to some extent).
}
\examples{

}
