% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{simulate_null.HiCglm}
\alias{simulate_null.HiCglm}
\title{Samples interactions based on background model.}
\usage{
\method{simulate_null}{HiCglm}(hic.glm, dpools = "all", N = 10)
}
\arguments{
\item{hic.glm}{HiCglm object}

\item{dpools}{character or vector, if character "all" then use all pooled diagonals in background model, otherwise use specified subset}

\item{N}{integer number of samples for every predictor value}
}
\value{
list of list of data frames with expected interactions, first list enumerates different maps (usually chromosomes), second list contains 2 elements refering to Y | X and X | Y models. Finally the last 2 lists contain data frames with simulated interactions.
}
\description{
Based on this method output one may plot expected vs observed plot of interactions for each diagonal pool (see examples).
}
\examples{
# first prepare HiCglm object
expected <- simulate_null(hicglm, dpools = 1:10)
observed <- lapply(hicglm$model[["18"]][as.character(1:10)], function(x) x$data)
observed <- do.call("rbind", observed)[c("val.x", "val.y", "dpool")]
# plot results: expected vs observed
ggplot(all.data, aes(x = val.x, y = val.y, color = type)) +
  geom_point(size = 0.3, alpha = 0.7) +
  facet_wrap(~ dpool, scales = "free", ncol = 3)

}
