% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sparse2interactions}
\alias{sparse2interactions}
\title{Convert sparse contact map to atomic interactions.}
\usage{
sparse2interactions(sparse.mtx)
}
\arguments{
\item{sparse.mtx}{data.frame Hi-C contact map in sparse format with mandatory columns i, j, val}
}
\value{
data.frame with 2 columns, which are i-th (x) and j-th (y) coordinates of every interaction.
}
\description{
Converts contact map in sparse data frame format into data frame with interaction, where one row is single interaction. This is usefull for bootstraping Hi-C interactions.
}
\examples{
# create data frame with artificial interactions, where val
# indicates total number of interactions between bins i and j
sparse.mtx <- data.frame(i = c(1,3,5,7,8,9,11), j = c(7,2,1,1,3,10,9), val = c(10,8,3,1,1,2,20))
print(sparse.mtx)
sparse2interactions(sparse.mtx)

}
\seealso{
\code{\link{dense2sparse}}
}
