% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{constructGLM}
\alias{constructGLM}
\title{Constructs GLM to model per-diagonal Hi-C contact dependancy}
\usage{
constructGLM(
  df,
  robust.nb = TRUE,
  overdisp.test.pval = 0.01,
  max.nobs = 20000,
  nrep = 10
)
}
\arguments{
\item{df}{data frame with predictor, response, outlier columns}

\item{robust.nb}{logical whther to use robust fitting procedure (see details)}

\item{overdisp.test.pval}{numeric significance threshold for testing ovedispersion}

\item{max.nobs}{numeric maximum number of observations (points), i.e. sample size to be taken for robust NB regression estimation (see details)}

\item{nrep}{numeric number of repetitions for subsampling (see details)}
}
\value{
object of class glm or MASS::glm.nb
}
\description{
Models Hi-C contacts using (robust) Negative Binomial (or Poisson when the data is underdispersed) regression. Given the fact that Hi-C data suffers from contact decay bias this method is intendent to model each diagonal separately. By default this function uses robust Negative Binomial regression to model interaction dependencies.
}
\details{
If \code{robust.nb} is true then this function uses robust Negative Binomial estimation method developed in \insertCite{aeberhard2014robust}{DIADEM}. This function uses the code of glmrob.nb function written by William Aeberhard, which is available at: https://github.com/williamaeberhard/glmrob.nb.

At first overdispersion test is performed to decide if Negative Binomial or Poisson regression should be used. If \code{robust.nb} is true the estimation may consume huge amounts of memory for large sample sizes (like for example 400000 points). In order to prevent that whenever the sample size exceeds \code{max.nobs} initial sample is subsampled to \code{max.nobs} size and model is estimated on subsample. This procedure is repeated \code{nrep} times and final parameter estimate equals average over subsampled estimates.
}
\references{
{
  \insertRef{aeberhard2014robust}{DIADEM}
}
}
\seealso{
\code{\link{glm}}, \code{\link[MASS]{glm.nb}} to see how GLM are constructed, \code{\link[AER]{dispersiontest}} to see how overdispersion is tested
}
