% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{decay_correlation.HiCcomparator}
\alias{decay_correlation.HiCcomparator}
\title{Calculates correlations between diagonals.}
\usage{
\method{decay_correlation}{HiCcomparator}(hic.comparator)
}
\arguments{
\item{hic.comparator}{object of type HiCcomparator}
}
\value{
dataframe with following columns: diagonal, pcc, pearson.pval, rho, spearman.pval, tau, kendall.pval, name which can be used to conveniently visualise dependancy between 2 Hi-C maps being compared (see examples)
}
\description{
Computes correlations (Pearson, Spearman, Kendall) and significances of corresponding diagonals between 2 Hi-C maps of HiCcomparator object.
}
\examples{
first create HiCcomparator object - see ?HiCcomparator for examples
library("ggplot2")
library("reshape2")
decay.cors <- decay_correlation(hic.comparator)
# wide to long
decay.cors.long <- reshape2::melt(decay.cors[c("name","diagonal","pcc","rho","tau")], id.vars = c("name","diagonal"), variable.name = "correlation", value.name = "coefficient")
# remove 0 diagonal (as it is non informative anyways) and illustrate results
ggplot(decay.cors.long[decay.cors.long$diagonal != 0,],
  aes(x = diagonal, y = coefficient, color = correlation)) +
 geom_point(size = 0.3) +
 facet_wrap(~ name, ncol = 1, scales = "free") +
 theme(legend.position = "bottom")

}
\seealso{
\code{\link{HiCcomparator}} on how to construct HiCcomparator object
}
