% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{differential_interactions.HiCcopula}
\alias{differential_interactions.HiCcopula}
\title{Finds significantly interacting rectangle-like regions.}
\usage{
\method{differential_interactions}{HiCcopula}(hic.copula,
  plot.models = TRUE)
}
\arguments{
\item{hic.copula}{object of class HiCcopula}

\item{plot.models}{logical if true then plot bilinear model fit for every matrix in hic.copula object; it will plot models for depleted and enriched models separately; if you want to save this results to file open device before calling this function (see for instance \code{\link{pdf}}) and close device after function call (see \code{\link{dev.off}})}
}
\value{
list with number of entries equal to hic.copula$names; each entry is a list with 2 elements: interacting.regions - data frame containing rows with rectangle like regions of significant interactions with coordinates n.cells (number of non zero cells inside rectangle), start.x, end.x, start.y, end.y, effect; connected.components list with cells comprising given connected component; connected components list is named list where each entry name is unique id, which can be mapped to row in interacting.regions (its row names)
}
\description{
This function works in 3 steps:
\itemize{
 \item{}{first it calculates differential map,}
 \item{}{then it takes negative log10 p-value vector of cells, sorts it, divides on negative and positive parts and lastly fits bilinear model to each of them}
 \item{}{finally it retains only those cells that are to the left (right) of intersection point of bilinear model in positive (negative) p-values vector and searches for connected components in both of them separately}
}
Fitting bilinear model is performed using \code{\link{best_fit_bilinear}} function, while for connected components search \code{\link{raster}} package is used. After detection of significantly interacting regions (connected components) one may further filter list to only retain those with number of non zero cells (n.cells column in interacting.regions data frame) larger than some threshold.
}
\examples{
# first create HiCcopula object - see ?HiCcopula for examples
di <- differential_interactions(hic.copula)
di18 <- di[["18"]]
# if you want to plot results create pvalue map (see hicdiff function) and from that dense map for some chromosome
plot_diff_map(dense)
# plot regions having at least 10 significant cells in connected component
plot_regions(di18[di18$n.cells >= 10,2:6], pal.colors = c("blue","red"))

}
\seealso{
\code{\link{best_fit_bilinear}} for fitting bilinear model, \code{\link{raster::raster}} and \code{\link{raster::clump}} for connected components search
}
