% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{HiCcomparator}
\alias{HiCcomparator}
\title{An S3 class to represent object for Hi-C maps comparisons.}
\usage{
HiCcomparator(path1, path2, tads = NULL, mtx.names = "all",
  which.tads = 4, do.pca = FALSE, zscore = FALSE, props = FALSE)
}
\arguments{
\item{path1}{character - path to npz file containing first set of Hi-C maps}

\item{path2}{character - path to npz file containing second set of Hi-C maps}

\item{tads}{list (optional), set of TADs as named list of data frames, each with at least start, end columns}

\item{mtx.names}{character vector with subset of Hi-C maps names to be selected for analysis, by default all matrices are used}

\item{which.tads}{numeric indicating what to do if no TADs are specified: 1 - determine TADs from first set of Hi-C maps, 2 - determine TADs from second set of Hi-C maps, 3 - determine from both sets and then take their intersection, 4 - do not determine TADs}

\item{do.pca}{logical whether to perform PCA for given maps and determine A/B compartments}
}
\value{
S3 object of class HiCcomparator
}
\description{
HiC comparator object stores Hi-C contact maps from 2 experiments and (optionally) TADs
and allows for convenient access to contact matrices, A/B compartments or TADs. HiCcomparator
is constructed from npz files containing Hi-C maps in python dict with numpy matrices.
Additionally TAD set may be given to HiCcomparator (as list of data frames, where data frames
names match those of Hi-C matrices names). One can also choose to determine TADs based on
given Hi-C contact maps - only first, only second or determine both and take intersecting
intervals between them.
}
\examples{
# get path of first sample maps
mtx1.fname <- system.file("extdata", "IMR90-MboI-1_40kb-raw.npz", package = "DIADEM", mustWork = TRUE)
# get path of second sample maps
mtx2.fname <- system.file("extdata", "MSC-HindIII-1_40kb-raw.npz", package = "DIADEM", mustWork = TRUE)
# get sample TADs
tads <- DIADEM::sample_tads[c("IMR90-MboI-1_40kb-raw", "MSC-HindIII-1_40kb-raw")]
# construct HiCcomparator object for chromosomes 18 and 19
hic.comparator <- HiCcomparator(mtx1.fname, mtx2.fname, tads, mtx.names = c("18","19"))
# plot A/B compartments for first and second map in chromosome 19
plot_pc_vector(hic.comparator$pc1.maps1[["19"]]) # first map
plot_pc_vector(hic.comparator$pc1.maps2[["19"]]) # second map

}
\seealso{
\code{\link{read_npz}} for reading npz files, \code{\link{do_pca}} on how A/B compartments are determined, \code{\link{map2tads}} how TADs are determined
}
