% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{HiCglm}
\alias{HiCglm}
\title{An S3 object to represent differential GLM Hi-C model.}
\usage{
HiCglm(
  hic.comparator,
  diag.frac = 0.5,
  alpha = 0.05,
  robust.nb = TRUE,
  remove.outliers = TRUE,
  outlier.weight = 0,
  ncores = 1,
  max.nobs = 10000,
  nrep = 20
)
}
\arguments{
\item{hic.comparator}{object of type HiCcomparator}

\item{diag.frac}{fraction of diagonals to use to fit models, by default fraction of chromsome length is used to indicate number of diagonals.}

\item{alpha}{significance (p-value) threshold to discard diagonals with insignificant Pearson correlation coefficient between X (contacts from contact map 1) and Y (corresponding contacts from contact map 2).}

\item{robust.nb}{logical whether to use robust Negative Binomial regression as main model (see details), true by default}

\item{remove.outliers}{logical only relevant if \code{robust.nb} is set to false - if true try to remove outliers (potential Differential Interactions) before model fitting (NB regression) using robust regression (IRLS) with bisquare weight function}

\item{outlier.weight}{numeric weight threshold to remove outliers, 0 by default}

\item{ncores}{numeric number of cores for parallel processing to speed up computations}

\item{max.nobs}{numeric maximum number of observations to be used for robust NB regression model estimation (see \code{\link{constructGLM}})}

\item{nrep}{number of repetitions to average parameter estimates during robust NB regression model estimation on large sample sizes (see \code{\link{constructGLM}})}
}
\value{
S3 object of class HiCglm
}
\description{
Models diagonal-wise dependencies between Hi-C data sets with GLM. Model is constructed as follows:
\itemize{
  \item{}{merge maps1 with maps2}
  \item{}{for each diagonal in diagonal pools}
  \itemize{
    \item{}{take all points from this diagonal, such that they are non zero in map1 (X) and non zero in map2 (Y)}
    \item{}{remove outliers using robust regression (only performed when \code{robust.nb} is false)}
    \item{}{model response (Y = f(X) or X = f(Y)) using (robust) Negative Binomial distribution}
  }
}
Before fitting the model it's recommended to first inspect correlations between analyzed Hi-C maps before fixing this variable. As the ratio of noise / signal in Hi-C data increases rapidly with decay it's unadvised to use all diagonals for modelling. The number of diagonals to be used will depend on chromosome length, resolution and data quality. One can retain uncorrelated diagonals (pools) by increasing significance threshold (\code{alpha}) for PCC significance.
}
\details{
When \code{robust.nb} is true, the model  is fitted using robust procedure developed and published by \insertCite{aeberhard2014robust}{DIADEM} - see \code{\link{constructGLM}} function for details.

The main assumption behind \code{HiCglm} is that given 2 Hi-C datasets (even very different cell lines) their diagonal-wise interaction profiles will be correlated, which should manifest in X,Y plots having funnel like shape (see \code{simulate_null} function). In this setting it is further assumed (based on observed X,Y dependencies) that potential Differential Interactions will behave like outliers thereby disobeying the funnel-like pattern. The use of robust Negative Binomial regression  diminishes the influence of outliers on model fitting and allows to capture “uncontaminated” (with no DIs) null model.
}
\examples{
# first create hiccomp (i.e. HiCcomparator object), then
hicglm <- HiCglm(hiccomp)

}
\references{
{
  \insertRef{aeberhard2014robust}{DIADEM}
}
}
\seealso{
\code{\link{HiCcomparator}} on how to construct HiCcomparator object and \code{\link{constructGLM}}, \code{\link[robustreg]{robustRegBS}} on how GLM fitting process
}
