% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{map2tads}
\alias{map2tads}
\title{Determines TAD boundaries using Insulation Score.}
\usage{
map2tads(
  dense.mtx,
  resolution = 40000,
  window.bp = 500 * 1000,
  delta.bp = 100 * 1000,
  without_unmappable = TRUE
)
}
\arguments{
\item{dense.mtx}{numeric matrix in dense format representing Hi-C contact map}

\item{resolution}{numeric resolution of Hi-C contact map in base pairs}

\item{window.bp}{numeric size of sliding window in base pairs}

\item{delta.bp}{numeic size of delta window in base pairs}
}
\value{
data frame with TAD positions containing start, end columns
}
\description{
For details on Insulation Score approach of TAD boundaries detection see Crane et al. 2015 "Condensin-driven remodelling of X chromosome topology during dosage compensation" methods section, paragraph TAD calling (insulation square analysis).
}
\examples{
# get Hi-C contact map
sparse.mtx <- DIADEM::sample_hic_maps[["MSC-HindIII-1_40kb-raw"]][["18"]]
dense.mtx <- sparse2dense(sparse.mtx[c("i","j","val")], N = 1952)
# get tads
tads <- map2tads(dense.mtx)
# plot results
plot_contact_map(dense.mtx)
plot_regions(tads)

}
